requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
joblib==1.1.0
numpy==1.22.2
psycopg2==2.9.3
python-crfsuite==0.9.7
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn-crfsuite==0.3.6
sqlparse==0.4.2
tabulate==0.8.9
threadpoolctl==3.1.0
tqdm==4.63.0
whitenoise==6.0.0
